requirements.txt
Fabric==1.10.2
paramiko==1.10.1
parse
pep8==1.7
pyyaml
boto3==1.3.1
